SatelliteDataSources

SatelliteDataSources is a pure Julia package built on top of Rasters.jl for reading and manipulating satellite imagery. Each AbstractSatellite provides a set of layers that can be conveniently read into either a Raster or RasterStack. Additionally, all AbstractSatellite types define a collection of sensor-specific information, such as digital number encoding, band wavelength, and band color.

Supported Sensors

Several popular satellites are already supported, and we provide a simple interface for adding more down the line.

SatelliteDataSources.AbstractSatellite — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

SatelliteDataSources.Landsat7 — Type

Implements the AbstractSatellite interface for Landsat 7.

Supported Layers: :B1, :B2, :B3, :B4, :B5, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :panchromatic, :thermal, :dilated_clouds, :clouds, :cloud_shadow, :snow, :water

SatelliteDataSources.Landsat8 — Type

Implements the AbstractSatellite interface for Landsat 8.

Supported Layers: :B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :panchromatic, :thermal1, :thermal2, :dilated_clouds, :clouds, :cloud_shadow, :snow, :water

SatelliteDataSources.Landsat9 — Type

Implements the AbstractSatellite interface for Landsat 9.

Supported Layers: :B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :panchromatic, :thermal1, :thermal2, :dilated_clouds, :clouds, :cloud_shadow, :snow, :water

SatelliteDataSources.Sentinel2 — Type

Implements the AbstractSatellite interface for Sentinel 2. The user must specify a resolution of eith 10, 20, or 60 meters.

Supported Layers (10m): :B02, :B03, :B04, :B08, :blue, :green, :red, :nir

Supported Layers (20m): :B02, :B03, :B04, :B05, :B06, :B07, :B8A, :B11, :B12, :blue, :green, :red, :nir, :swir1, :swir2, :cloud_shadow, :clouds_med, :clouds_high, :cirrus, :vegetation, :soil, :water, :snow

Supported Layers (60m): :B01, :B02, :B03, :B04, :B05, :B06, :B07, :B8A, :B09, :B11, :B12, :blue, :green, :red, :nir, :swir1, :swir2, :cloud_shadow, :clouds_med, :clouds_high, :cirrus, :vegetation, :soil, :water, :snow

User Interface

SatelliteDataSources.bands — Function

bands(::Type{AbstractSatellite})

Return the band names in order from shortest to longest wavelength.

SatelliteDataSources.layers — Function

layers(::Type{AbstractSatellite})

Return the names of all layers available for the given sensor.

SatelliteDataSources.wavelengths — Function

wavelengths(::Type{AbstractSatellite})

Return the central wavelengths for all bands in order from shortest to longest.

SatelliteDataSources.wavelength — Function

wavelength(::Type{AbstractSatellite}, band::Symbol)

Return the central wavelength for the corresponding band.

SatelliteDataSources.blue_band — Function

blue_band(::Type{AbstractSatellite})

Return the blue band for the given sensor.

SatelliteDataSources.green_band — Function

green_band(::Type{AbstractSatellite})

Return the green band for the given sensor.

SatelliteDataSources.red_band — Function

red_band(::Type{AbstractSatellite})

Return the red band for the given sensor.

SatelliteDataSources.nir_band — Function

nir_band(::Type{AbstractSatellite})

Return the nir band for the given sensor.

SatelliteDataSources.swir1_band — Function

swir1_band(::Type{AbstractSatellite})

Return the swir1 band for the given sensor.

SatelliteDataSources.swir2_band — Function

swir2_band(::Type{AbstractSatellite})

Return the swir2 band for the given sensor.

SatelliteDataSources.dn_scale — Function

dn_scale(::Type{AbstractSatellite}, layer::Symbol)

Return the scale factor applied to digital numbers.

SatelliteDataSources.dn_offset — Function

dn_offset(::Type{AbstractSatellite}, layer::Symbol)

Return the offset factor applied to digital numbers.

SatelliteDataSources.decode — Function

decode(s::Type{AbstractSatellite}, raster::Rasters.AbstractRaster)
decode(s::Type{AbstractSatellite}, raster::Rasters.AbstractRasterStack)

Decode the Digital Number (DN) values in the given raster(s). Typically, the decoded values will be in either reflectance (visual bands) or Kelvin (thermal bands).

Parameters

s: The AbstractSatellite that produced the raster(s) in question.

raster: Either a Rasters.Raster or Rasters.RasterStack to be decoded.

SatelliteDataSources.encode — Function

encode(s::Type{AbstractSatellite}, raster::Rasters.AbstractRaster; encoding_type=UInt16, missingval=0x0000)
encode(s::Type{AbstractSatellite}, raster::Rasters.AbstractRasterStack; kwargs...)

Encode the provided raster(s) to Digital Numbers (DN).

Parameters

s: The AbstractSatellite that produced the raster(s) in question.

raster: Either a Rasters.Raster or Rasters.RasterStack to be encoded.

encoding_type: The Julia type to use for storing the encoded values (default = UInt16).

missingval: The value to denote missing values (default = 0x0000).

Rasters.Raster — Type

Raster(s::Type{AbstractSatellite}, dir::String, layer::Symbol; kwargs...)

Read a single layer into a Rasters.Raster.

Parameters

s: The AbstractSatellite that produced the raster(s) in question.

dir: The directory containing the satellite layers.

layer: The layer to be read. See layers(s) for a list of available layers for sensor s.

Layer Sources

A sensor's layers can come form a variety of sources, including single-band rasters, multiband rasters, and bit-fields. However, we do not want to expose these particulars to the end user. Thus, we rely on several Julia types to represent this abstraction. Each of these types are sub-types of AbstractLayerSource and store the necessary information to retrieve the corresponding layer.

SatelliteDataSources.AbstractLayerSource — Type

Super type of all layer sources.

SatelliteDataSources.File — Type

A layer corresponding to a file.

Commonly used for sensors which store their bands as individual rasters.

SatelliteDataSources.Band — Type

A layer corresponding to a particular band in a multi-band file.

Commonly used for sensors which store their bands in a single multi-band raster.

SatelliteDataSources.MaskValue — Type

A layer corresponding to a specific value in a single-band file.

Sentinel 2 uses this format to encode segmentation masks in the SCL file.

SatelliteDataSources.BitField — Type

A layer corresponding to a particular bit in a single-band file.

Landsat uses this format to encode segmentation masks within the QA file.

SatelliteDataSources.layer_source — Function

layer_source(::Type{AbstractSatellite}, layer::Symbol)

Retrieve the AbstractLayerSource for the given layer and sensor type.

Example

using Rasters, SatelliteDataSources

# Path to Satellite Product
src = "data/LC08_L2SP_043024_20200802_20200914_02_T1"

# Load the Blue, Green, Red, and NIR Bands
stack = RasterStack(Landsat8, src, [:blue, :green, :red, :nir], lazy=true)

# Mask Clouds
cloud_mask = Raster(Landsat8, src, :clouds) 
shadow_mask = Raster(Landsat8, src, :cloud_shadow) 
raster_mask = .!(boolmask(cloud_mask) .|| boolmask(shadow_mask))
masked_stack = mask(stack, with=raster_mask)

# Save Processed Data as a Multiband Raster
masked_raster = Raster(masked_stack)
write("data/masked_bands.tif", masked_raster)

Index

Documentation for SatelliteDataSources.

Rasters.Raster

SatelliteDataSources.AbstractLayerSource

SatelliteDataSources.AbstractSatellite

SatelliteDataSources.Band

SatelliteDataSources.BitField

SatelliteDataSources.File

SatelliteDataSources.Landsat7

SatelliteDataSources.Landsat8

SatelliteDataSources.Landsat9

SatelliteDataSources.MaskValue

SatelliteDataSources.Sentinel2

SatelliteDataSources.bands

SatelliteDataSources.blue_band

SatelliteDataSources.decode

SatelliteDataSources.dn_offset

SatelliteDataSources.dn_scale

SatelliteDataSources.encode

SatelliteDataSources.green_band

SatelliteDataSources.layer_source

SatelliteDataSources.layers

SatelliteDataSources.nir_band

SatelliteDataSources.red_band

SatelliteDataSources.swir1_band

SatelliteDataSources.swir2_band

SatelliteDataSources.wavelength

SatelliteDataSources.wavelengths