SatelliteDataSources

Documentation for SatelliteDataSources.

SatelliteDataSources.AbstractSatellite

SatelliteDataSources.Landsat8

SatelliteDataSources.Landsat9

SatelliteDataSources.Sentinel2

SatelliteDataSources.bandnames

SatelliteDataSources.blue_band

SatelliteDataSources.dn_offset

SatelliteDataSources.dn_scale

SatelliteDataSources.getlayers

SatelliteDataSources.green_band

SatelliteDataSources.layernames

SatelliteDataSources.nir_band

SatelliteDataSources.parse_layers

SatelliteDataSources.red_band

SatelliteDataSources.swir1_band

SatelliteDataSources.swir2_band

SatelliteDataSources.wavelength

SatelliteDataSources.wavelengths

SatelliteDataSources.AbstractSatellite — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

SatelliteDataSources.Landsat8 — Type

Implements the AbstractSatellite interface for Landsat 8.

Supported layers are: (:B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :QA, :bands).

SatelliteDataSources.Landsat9 — Type

Implements the AbstractSatellite interface for Landsat 9.

Supported layers are: (:B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :QA).

SatelliteDataSources.Sentinel2 — Type

Implements the AbstractSatellite interface for Sentinel 2.

SatelliteDataSources.bandnames — Function

bandnames(::Type{AbstractSatellite})

Return the band names in order from shortest to longest wavelength.

SatelliteDataSources.blue_band — Method

blue_band(::Type{AbstractSatellite})

Return the blue band for the given sensor.

SatelliteDataSources.dn_offset — Method

dn_offset(::Type{AbstractSatellite})

Return the offset applied to digital numbers.

SatelliteDataSources.dn_scale — Method

dn_scale(::Type{AbstractSatellite})

Return the scaling factor applied to digital numbers.

SatelliteDataSources.getlayers — Method

getlayers(::Type{T}, dir::String; layers=bandnames(T)) where {T <: AbstractSatellite}

Retrieve the requested layers in the provided directory. Returns a dictionary where layer names are keys and filenames are values.

SatelliteDataSources.green_band — Method

green_band(::Type{AbstractSatellite})

Return the green band for the given sensor.

SatelliteDataSources.layernames — Function

layernames(::Type{AbstractSatellite})

Return the names of all layers available for the given sensor.

SatelliteDataSources.nir_band — Method

nir_band(::Type{AbstractSatellite})

Return the nir band for the given sensor.

SatelliteDataSources.parse_layers — Function

parse_layers(::Type{AbstractSatellite}, dir::String)

Retrieve all layers in the provided directory. Returns a dictionary where layer names are keys and filenames are values.

SatelliteDataSources.red_band — Method

red_band(::Type{AbstractSatellite})

Return the red band for the given sensor.

SatelliteDataSources.swir1_band — Method

swir1_band(::Type{AbstractSatellite})

Return the swir1 band for the given sensor.

SatelliteDataSources.swir2_band — Method

swir2_band(::Type{AbstractSatellite})

Return the swir2 band for the given sensor.

SatelliteDataSources.wavelength — Method

wavelength(::Type{AbstractSatellite}, band::Symbol)

Return the central wavelength for the corresponding band.

SatelliteDataSources.wavelengths — Function

wavelengths(::Type{AbstractSatellite})

Return the central wavelengths for all bands in order from shortest to longest.