SatelliteTypes

Documentation for SatelliteTypes.

SatelliteTypes.AbstractBandset

SatelliteTypes.Landsat8

SatelliteTypes.Sentinel2

SatelliteTypes.bandnames

SatelliteTypes.blue_band

SatelliteTypes.getlayers

SatelliteTypes.green_band

SatelliteTypes.layernames

SatelliteTypes.nir_band

SatelliteTypes.parse_layers

SatelliteTypes.red_band

SatelliteTypes.swir1_band

SatelliteTypes.swir2_band

SatelliteTypes.wavelength

SatelliteTypes.wavelengths

SatelliteTypes.AbstractBandset — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

SatelliteTypes.Landsat8 — Type

Implements the AbstractBandset interface for Landsat 8.

Supported layers are: (:B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :QA, :bands).

SatelliteTypes.Sentinel2 — Type

Implements the AbstractBandset interface for Sentinel 2.

SatelliteTypes.bandnames — Method

bandnames(::Type{AbstractBandset})

Return the band names in order from shortest to longest wavelength.

SatelliteTypes.blue_band — Method

blue_band(::Type{AbstractBandset})

Return the blue band for the given sensor.

SatelliteTypes.getlayers — Method

getlayers(::Type{T}, dir::String; layers=bandnames(T)) where {T <: AbstractBandset}

Retrieve the requested layers in the provided directory. Returns a dictionary where layer names are keys and filenames are values.

SatelliteTypes.green_band — Method

green_band(::Type{AbstractBandset})

Return the green band for the given sensor.

SatelliteTypes.layernames — Method

layernames(::Type{AbstractBandset})

Return the names of all layers available for the given sensor.

SatelliteTypes.nir_band — Method

nir_band(::Type{AbstractBandset})

Return the nir band for the given sensor.

SatelliteTypes.parse_layers — Method

parse_layers(::Type{AbstractBandset}, dir::String)

Retrieve all layers in the provided directory. Returns a dictionary where layer names are keys and filenames are values.

SatelliteTypes.red_band — Method

red_band(::Type{AbstractBandset})

Return the red band for the given sensor.

SatelliteTypes.swir1_band — Method

swir1_band(::Type{AbstractBandset})

Return the swir1 band for the given sensor.

SatelliteTypes.swir2_band — Method

swir2_band(::Type{AbstractBandset})

Return the swir2 band for the given sensor.

SatelliteTypes.wavelength — Method

wavelength(::Type{AbstractBandset}, band::Symbol)

Return the central wavelength for the corresponding band.

SatelliteTypes.wavelengths — Method

wavelengths(::Type{AbstractBandset})

Return the central wavelengths for all bands in order from shortest to longest.