SatelliteDataSources

Documentation for SatelliteDataSources.

SatelliteDataSources.AbstractLayerSource

SatelliteDataSources.AbstractSatellite

SatelliteDataSources.Band

SatelliteDataSources.BitField

SatelliteDataSources.File

SatelliteDataSources.Landsat7

SatelliteDataSources.Landsat8

SatelliteDataSources.Landsat9

SatelliteDataSources.MaskValue

SatelliteDataSources.Sentinel2

SatelliteDataSources.bands

SatelliteDataSources.blue_band

SatelliteDataSources.dn_offset

SatelliteDataSources.dn_scale

SatelliteDataSources.green_band

SatelliteDataSources.layer_source

SatelliteDataSources.layers

SatelliteDataSources.nir_band

SatelliteDataSources.red_band

SatelliteDataSources.swir1_band

SatelliteDataSources.swir2_band

SatelliteDataSources.wavelength

SatelliteDataSources.wavelengths

SatelliteDataSources.AbstractLayerSource — Type

Super type of all concrete layer sources.

SatelliteDataSources.AbstractSatellite — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

SatelliteDataSources.Band — Type

Represents a layer corresponding to a particular band in a multi-band file. Commonly used for sensors which store their bands as a single multi-band file.

SatelliteDataSources.BitField — Type

Represents a layer corresponding to a particular bit in a single-band file. Commonly used for QA and scene classification masks.

SatelliteDataSources.File — Type

Represents a layer corresponding to a file. Commonly used for sensors which store their bands in individual files.

SatelliteDataSources.Landsat7 — Type

Implements the AbstractSatellite interface for Landsat 8.

Supported Layers: :B1, :B2, :B3, :B4, :B5, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :panchromatic, :thermal, :dilatedclouds, :clouds, :cloudshadow, :snow, :water

SatelliteDataSources.Landsat8 — Type

Implements the AbstractSatellite interface for Landsat 8.

Supported Layers: :B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :panchromatic, :thermal1, :thermal2, :dilatedclouds, :clouds, :cloudshadow, :snow, :water.

SatelliteDataSources.Landsat9 — Type

Implements the AbstractSatellite interface for Landsat 9.

Supported Layers: :B1, :B2, :B3, :B4, :B5, :B6, :B7, :blue, :green, :red, :nir, :swir1, :swir2, :panchromatic, :thermal1, :thermal2, :dilatedclouds, :clouds, :cloudshadow, :snow, :water.

SatelliteDataSources.MaskValue — Type

Represents a layer corresponding to a particular bit in a single-band file. Commonly used for QA and scene classification masks.

SatelliteDataSources.Sentinel2 — Type

Implements the AbstractSatellite interface for Sentinel 2. The user must specify a resolution of eith 10, 20, or 60 meters.

Sentinel2{10} Layers: :B02, :B03, :B04, :B08, :blue, :green, :red, :nir

Sentinel2{20} Layers: :B02, :B03, :B04, :B05, :B06, :B07, :B8A, :B11, :B12, :blue, :green, :red, :nir, :swir1, :swir2, :cloudshadow, :cloudsmed, :clouds_high, :cirrus, :vegetation, :soil, :water, :snow

Sentinel2{60} Layers: :B01, :B02, :B03, :B04, :B05, :B06, :B07, :B8A, :B09, :B11, :B12, :blue, :green, :red, :nir, :swir1, :swir2, :cloudshadow, :cloudsmed, :clouds_high, :cirrus, :vegetation, :soil, :water, :snow

SatelliteDataSources.bands — Function

bands(::Type{AbstractSatellite})

Return the band names in order from shortest to longest wavelength.

SatelliteDataSources.blue_band — Method

blue_band(::Type{AbstractSatellite})

Return the blue band for the given sensor.

SatelliteDataSources.dn_offset — Method

dn_offset(::Type{AbstractSatellite}, layer::Symbol)

Return the offset factor applied to digital numbers.

SatelliteDataSources.dn_scale — Method

dn_scale(::Type{AbstractSatellite}, layer::Symbol)

Return the scale factor applied to digital numbers.

SatelliteDataSources.green_band — Method

green_band(::Type{AbstractSatellite})

Return the green band for the given sensor.

SatelliteDataSources.layer_source — Function

layer_source(::Type{AbstractSatellite}, layer::Symbol)

Retrieve the AbstractLayerSource for the given layer and sensor type.

SatelliteDataSources.layers — Function

layers(::Type{AbstractSatellite})

Return the names of all layers available for the given sensor.

SatelliteDataSources.nir_band — Method

nir_band(::Type{AbstractSatellite})

Return the nir band for the given sensor.

SatelliteDataSources.red_band — Method

red_band(::Type{AbstractSatellite})

Return the red band for the given sensor.

SatelliteDataSources.swir1_band — Method

swir1_band(::Type{AbstractSatellite})

Return the swir1 band for the given sensor.

SatelliteDataSources.swir2_band — Method

swir2_band(::Type{AbstractSatellite})

Return the swir2 band for the given sensor.

SatelliteDataSources.wavelength — Method

wavelength(::Type{AbstractSatellite}, band::Symbol)

Return the central wavelength for the corresponding band.

SatelliteDataSources.wavelengths — Function

wavelengths(::Type{AbstractSatellite})

Return the central wavelengths for all bands in order from shortest to longest.